requirements.txt
wheel
numpy
seaborn
scipy
matplotlib
tensorflow-gpu==1.15.0
sklearn
pexpect
pyyaml
https://github.com/dkollective/djx/tarball/mlflow